requirements.txt
argon2-cffi==18.1.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
Django==2.0
django-debug-toolbar==1.9.1
django-widget-tweaks==1.4.1
Faker==0.8.10
get==0.0.39
heroku==0.1.4
idna==2.6
Pillow==5.0.0
post==0.0.26
public==0.0.65
pycparser==2.18
python-dateutil==1.5
pytz==2017.3
query-string==0.0.28
request==0.0.26
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.1
urllib3==1.22
